+++ examples/custom_checkers/e9989_pycodestyle/e131_no_error.py
@@ -1,0 +1,5 @@
+my_dict = {
+    "key": "value",
+    "long": "the quick brown fox jumps over the "
+            "lazy dog",
+}